backend/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework>=3.12.4
djangorestframework-simplejwt>=4.7.2
django-filter>=2.4.0
djoser>=2.1.0
Pillow>=8.2.0
django-import-export>=2.7.0
reportlab
gunicorn>=20.0
